+++ src/untar/script.sh
@@ -4,7 +4,7 @@ set -eo pipefail
 
 extra_args=()
 
-TMPDIR=$(mktemp -d "$meta_temp_dir/$meta_functionality_name-XXXXXX")
+TMPDIR=$(mktemp -d "$meta_temp_dir/$meta_name-XXXXXX")
 function clean_up {
   [[ -d "$TMPDIR" ]] && rm -r "$TMPDIR"
 }