+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## MINOR CHANGES
 
-* `concat_text`: Allow input files to be a mixed bag of gzipped and plain text files (PR #xxx).
-* `concat_text`: Stream concatenated files to gzip (PR #xxx).
+* `concat_text`: Allow input files to be a mixed bag of gzipped and plain text files (PR #12).
+* `concat_text`: Stream concatenated files to gzip (PR #12).
 
 # craftbox 0.2.0